% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emoji_masterlist.R
\docType{data}
\name{emoji_dict}
\alias{emoji_dict}
\title{Emoji Master List}
\format{
A named list of emoji names and corresponding Unicode symbols.
}
\usage{
data(emoji_dict)
}
\description{
A list of emoji Unicode values with human-friendly names.
}
\examples{
data(emoji_dict)
head(emoji_dict)
}
\keyword{emoji}
