% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emoji_list_subset.R
\name{emoji_list_subset}
\alias{emoji_list_subset}
\title{List Available Emojis}
\usage{
emoji_list_subset()
}
\value{
A named character vector of emojis.
}
\description{
Returns a named list of all available emojis in the emoji dictionary.
}
