% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emoji_shortcut.R
\name{emoji_shortcut}
\alias{emoji_shortcut}
\title{Emoji Shortcut Function}
\usage{
emoji_shortcut()
}
\value{
Opens addin popup for emoji selection.
}
\description{
Designed to be bound to a keyboard shortcut.
}
\examples{
# emoji_shortcut()
}
