% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_emoji.R
\name{git_emoji}
\alias{git_emoji}
\title{Emoji for Git Commit}
\usage{
git_emoji(name)
}
\arguments{
\item{name}{The emoji name.}
}
\value{
Emoji string for use in git commit messages.
}
\description{
Get emoji for Git commit messages.
}
\examples{
# git_emoji("fire")
# git_emoji("heart")
}
