% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_emoji.R
\name{insert_emoji}
\alias{insert_emoji}
\title{Insert Emoji into R Script}
\usage{
insert_emoji(name)
}
\arguments{
\item{name}{The name of the emoji (e.g. "smile").}
}
\value{
Inserts the emoji at cursor in RStudio.
}
\description{
Inserts an emoji by name from the emoji_dict.
}
\examples{
# insert_emoji("start")
}
