% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_emoji_addin.R
\name{insert_emoji_addin}
\alias{insert_emoji_addin}
\title{Insert Emoji Addin for RStudio}
\usage{
insert_emoji_addin()
}
\value{
Inserts the selected emoji at the cursor position in the active R script or R Markdown document.
}
\description{
This RStudio addin allows users to easily insert emojis by name into their R scripts, R Markdown, or Quarto documents.
A popup will appear, prompting the user to enter the name of an emoji. If the emoji name is valid, the corresponding emoji will be inserted at the cursor's position in the active editor.
}
\examples{
# Use this addin by running insert_emoji_addin() from the RStudio Addins menu.
# A prompt will appear where you can type an emoji name like "smile" or "heart" to insert it.
}
