% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shiny_emoji.R
\name{shiny_emoji}
\alias{shiny_emoji}
\title{Insert Emoji Addin for RStudio (with Searchable Dropdown)}
\usage{
shiny_emoji()
}
\value{
Inserts the selected emoji at the cursor position in the active R script or R Markdown document.
}
\description{
This RStudio addin allows users to easily insert emojis by name into their R scripts, R Markdown, or Quarto documents.
A Shiny-based GUI with a searchable dropdown will appear, allowing the user to select an emoji.
Once an emoji is selected, it will be inserted at the cursor's position in the active editor.
}
\examples{
# Use this addin by running insert_emoji_addin() from the RStudio Addins menu.
}
