% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Emoji_list.R
\name{emoji_list}
\alias{emoji_list}
\title{Print list of All Available Emoji names and actual emojis in the console}
\usage{
emoji_list()
}
\value{
prints the name of the emoji and the emoji image in the console
}
\description{
Prints all emoji names with their corresponding emoji symbols from the master list.
}
\examples{
emoji_list()

}
