% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_emoji.R
\name{git_emoji}
\alias{git_emoji}
\title{Emoji for Git Commit}
\usage{
git_emoji(name)
}
\arguments{
\item{name}{A character vector of emoji names (e.g., \code{"fire"}, \code{"heart"}, or country names).}
}
\value{
A named character vector of emojis.
}
\description{
Get emoji(s) for Git commit messages by name.
}
\examples{
git_emoji("fire")
git_emoji(c("fire", "heart"))
try(git_emoji(c("Ethiopia", "UK", "Kenya")))
}
