% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_emoji.R
\name{insert_emoji}
\alias{insert_emoji}
\title{Insert Emoji into R Script}
\usage{
insert_emoji(name, default = NULL)
}
\arguments{
\item{name}{The name of the emoji (case-insensitive).}

\item{default}{A default emoji to use if the name is not found.}
}
\value{
The emoji used (invisibly).
}
\description{
Inserts an emoji by name from the emoji_dict into RStudio.
}
