% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_emoji_addin.R
\name{insert_emoji_addin}
\alias{insert_emoji_addin}
\title{Insert Emoji Addin for RStudio}
\usage{
insert_emoji_addin(test_input = NULL)
}
\arguments{
\item{test_input}{For internal testing. If supplied, this will be used instead of prompting the user.}
}
\value{
Inserted emoji or NULL
}
\description{
Insert Emoji Addin for RStudio
}
